Dependency Updates
Both for the Node and Rust sides of things! Testing on Windows this time :)

Note: I got it to work on my Chromebook (Ubuntu now, super awesome setup!), but I ran out of storage real fast with my small 16 GB for the whole laptop, so I didn't have much available to try and build an entire version of the app, haha. It ran great with webkit2gtk though, so that's amazing! Can't wait to see what WebView2 will be like over here on Windows.
Note 2: Started this commit, then realized I should make a build for the previous release for Windows, since I had only done one for macOS at that point. Worked like a charm on Windows too!