New! Create-Tauri-App; 0.2.0
Looked fun to check things out with Tauri again, and I'm glad I did! It makes a lot more sense now that I've been working with TypeScript. These two could make a great combo for something together.

Before trying anything new, I wanted to upgrade my preexisting demo to the latest version of Tauri. So, I essentially ran-through `npx create-tauri-app` from scratch, replaced all of the files from this project with those results, diffed out what changed, and re-added my original source files as needed.

Now I think I have a better understanding as to how Tauri works. I think the base implementations are part of the Rust-backed 'backend' of the app, which is the full native code. This is interacted with via IPC in the front-end browser view. When you import modules from the `@tauri-apps` JS APIs, you are simply importing functions which wrap over these IPC calls, and those wrappers are what are bundled with your front-end code, when you use something like Vite. So, in theory, you could in fact (I think?) load those from a CDN if you were weird (me, I'd be up to demo it out :P), if you didn't want to use a bundler or something. So this could be neat if you have a Service Worker setup and Import Map module resolution setup. Not sure why you'd have to need that if you are building for Tauri though, as it will include all of your sources in the app build anyways, so, meh.

Another really neat thing you could do with this tooling: building the outputs for your app using GitHub Actions! This is something new to try for me, so I didn't want to bunch that up with these initial updates to the demo.

https://github.com/tauri-apps/tauri-action

This solves a bunch of the issues in regards to needing a machine to provide for those platforms, which I really like. Not to say that this doesn't mean you shouldn't demo and test your app on those platforms, but at least you can provide something to work with over there, right out of the box, even if it might be out of your reach in terms of hardware or OS limitations or something.

I absolutely like this much better than Electron now; Other than maybe needing to learn more about using Electron itself, this is definitely the route I'd choose for this kind of thing, rather than with Electron anymore. Electron is great for what it can do, but it does make it very bloated because of that. And I think that bloated nature is what's holding it back from being able to progress into something that people will want to embrace in the future. It set the example for great things ahead, and now Tauri is the next in line to help with that. Now I'm curious as to what comes after Tauri, ooo!

It would be great down the line if something like this became standardized at the OS level, that seems fairly reasonable to me. If every (fine, most?) OSes have a built-in webview of some sort, it would be great for apps to target that universally, and to provide cross-platform IPC native system-calls like this, rather than needing to build those into the app itself. Maybe this is already a methodology; sounds like something that could be in production in the world today, and I just don't know about it yet.